v20.0: Hardening complete - all type-check & build pass
✅ Server type-check: 0 errors (fixed bigint, PreparedQuery, index signature issues)
✅ Client type-check: 0 errors (fixed tsconfig.json includes, lazyWithRetry types)
✅ Production build: succeeds (vite build works)
✅ SQL queries: COUNT(*) cast to integer for pgtyped
✅ pgtyped types: regenerated with correct types
✅ lint: all prettier/eslint errors fixed

Server fixes (NO CI weakening!):
- articlesRepository: fixed Article type mappings
- blackBookRepository: fixed PreparedQuery type casts
- communicationsRepository: fixed JSON.parse type guards
- documentsRepository: fixed PreparedQuery return types
- statsRepository: fixed bigint vs number casts
- tsconfig.server.json: strict mode MAINTAINED

Client fixes:
- tsconfig.json: fixed includes to exclude server files
- lazyWithRetry: fixed ComponentType return types
- App.tsx: fixed lazy loading type compatibility
- Added Vite types for __APP_VERSION__, __BUILD_DATE__

+++ packages/db/src/queries/admin.sql
@@ -6,8 +6,8 @@ SELECT
 
 /* @name getEntityAndDocumentCounts */
 SELECT 
-  (SELECT COUNT(*) FROM entities) as entities,
-  (SELECT COUNT(*) FROM documents) as documents;
+  (SELECT COUNT(*)::integer FROM entities) as entities,
+  (SELECT COUNT(*)::integer FROM documents) as documents;
 
 /* @name listUsers */
 SELECT id, username, email, role, created_at as "createdAt", last_active as "lastActive"
